git-svn-id: http://grubutils.googlecode.com/svn@45 d95a8657-0e37-07c3-0af6-720afd5d0260

+++ trunk/src/include/grub4dos.h
@@ -205,7 +205,7 @@ typedef enum
 #define open ((int (*)(char *))((*(int **)0x8300)[26]))
 #define read ((unsigned long long (*)(unsigned long long, unsigned long long, unsigned long))((*(int **)0x8300)[27]))
 #define close ((void (*)(void))((*(int **)0x8300)[28]))
-#define devread ((int (*)(unsigned long sector, unsigned long byte_offset, unsigned long long byte_len, unsigned long long buf, unsigned long write))((*(int **)0x8300)[32]))
+#define devread ((int (*)(unsigned long long sector, unsigned long byte_offset, unsigned long long byte_len, unsigned long long buf, unsigned long write))((*(int **)0x8300)[32]))
 /*
  * int
  * devwrite (unsigned long sector, unsigned long sector_count, char *buf)
@@ -238,9 +238,9 @@ devread (unsigned long drive, unsigned long sector, unsigned long byte_offset, u
 #define unicode_to_utf8 ((void (*)(unsigned short *, unsigned char *, unsigned long))((*(int **)0x8300)[67]))
 /*
 int
-rawread (unsigned long drive, unsigned long sector, unsigned long byte_offset, unsigned long byte_len, unsigned long long buf, unsigned long write)
+rawread (unsigned long drive, unsigned long long sector, unsigned long byte_offset, unsigned long byte_len, unsigned long long buf, unsigned long write)
 */
-#define rawread ((int (*)(unsigned long, unsigned long, unsigned long, unsigned long long, unsigned long long, unsigned long))((*(int **)0x8300)[68]))
+#define rawread ((int (*)(unsigned long, unsigned long long, unsigned long, unsigned long long, unsigned long long, unsigned long))((*(int **)0x8300)[68]))
 /*
 int
 rawwrite (unsigned long drive, unsigned long sector, char *buf)
